It is calculator but have some problem
in the code in has shown a faulty calculator. I am a beginner at coding and made my first calculator. This calculator works perfectly but there is a problem. It gives the wrong answer on some specific equation. It's a very easy and simple program. It is my first exercise done by python programming. I made this using only "if", "else" and "elif" statements. I know there might be more paths to make this code but I have tried my best to make it. This code gives wrong answers to the equation of "45*3", "56+9", "56/6". For the first equation, it gives the "5555". And for the second it gives "77". And for the last one, it gives "4". You can see it clearly gives the wrong answer to these equations. But except for these three equations, it perfectly calculates everything. Also, thanks to you for checking my program.