A Guessing Game
It is a guessing game. Made by simple python program. It uses while to perform. In the program, you have to guess a number. The program will tell you if you have to think higher number or lower number. Also you only 9 guesses. You run out of guess you lose. So be careful in guessing. This is my second program.